requirements.txt
celery django-celery-results 
django-celery-beat 
flower
pika
redis
django-anymail
